main.py
### Updated Dummy Code Snippets

#### `main.py`
```python
from flask import Flask, render_template, request
from socketio import Server, WSGIApp
import os
from utils.recorder import save_chat_log, save_file
from utils.ai_chat import get_ai_response

app = Flask(__name__)
sio = Server(async_mode='threading')
app.wsgi_app = WSGIApp(sio, app.wsgi_app)

@app.route('/')
def index():
    return render_template('index.html')

@sio.event
def connect(sid, environ):
    print(f"User {sid} connected")

@sio.event
def message(sid, data):
    msg = data.get('text', '')
    save_chat_log(msg)
    ai_reply = get_ai_response(msg)  # AI response
    sio.emit('message', {'text': msg, 'ai_reply': ai_reply, 'sender': sid})
    print(f"Message: {msg}")

@sio.event
def file_upload(sid, data):
    filename = data.get('filename')
    content = data.get('content', 'Sample file content')
    save_file(filename, content)
    sio.emit('file_shared', {'filename': filename, 'sender': sid})

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5000)