requirements.txt
Flask==2.3.2
python-socketio==5.11.0
psycopg2-binary==2.9.6
numpy==1.24.3
google-generativeai==0.3.2  # Google Gemini API